@@ -1,22 +1,26 @@
 apply plugin: 'com.android.application'
-apply plugin: 'org.greenrobot.greendao'
 
 android {
-    compileSdkVersion 26
-    buildToolsVersion '26.0.2'
+    compileSdkVersion 28
+    buildToolsVersion '28.0.3'
 //    useLibrary "org.apache.http.legacy"
 
     defaultConfig {
         applicationId "me.gensh.helloustb"
         minSdkVersion 16
-        targetSdkVersion 26
-        versionCode 113
-        versionName "2.1.1"
-        buildConfigField "String", "APP_BUILD_TIME", "\"Mon, 30 October. 2017 \""   // build time
-        testInstrumentationRunner "android.support.test.runner.AndroidJUnitRunner"
+        targetSdkVersion 28
+        versionCode 114
+        versionName "2.2.0"
+        buildConfigField "long", "APP_BUILD_TIME", System.currentTimeMillis() + "L" // build time
+        testInstrumentationRunner "androidx.test.runner.AndroidJUnitRunner"
         externalNativeBuild {
             cmake {
                 cppFlags ""
+            }
+        }
+        javaCompileOptions {
+            annotationProcessorOptions {
+                arguments = ["room.schemaLocation": "$projectDir/schemas".toString()]
             }
         }
 //        jackOptions {  //for java8
