@@ -6,7 +6,7 @@ import android.content.Intent;
 import android.content.Context;
 import android.net.Uri;
 import android.os.Build;
-import android.support.v4.content.FileProvider;
+import androidx.core.content.FileProvider;
 import android.widget.Toast;
 
 import java.io.File;
