@@ -19,8 +19,9 @@ import android.os.Environment;
 import android.os.Parcelable;
 import android.provider.DocumentsContract;
 import android.provider.MediaStore;
-import android.support.annotation.ColorInt;
 import android.util.TypedValue;
+
+import androidx.annotation.ColorInt;
 
 import java.util.List;
 import java.util.Stack;
