@@ -1,8 +1,9 @@
 package me.gensh.views;
 
 import android.content.Context;
-import android.support.v7.widget.AppCompatImageView;
 import android.util.AttributeSet;
+
+import androidx.appcompat.widget.AppCompatImageView;
 
 /**
  * Created by gensh on 2017/9/5.
